// packages/sdk-ts/src/schemas/generated.ts
export interface paths {
// ...
    "/rss/pickup-job/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** RssGetOnePickupJob */
        get: operations["getRssPickupjobGet"];
        put?: never;
        /** RssGetOnePickupJob */
        post: operations["postRssPickupjobGet"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/rss/pickup-jobs/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** RssGetPickupJobs */
        get: operations["getRssPickupjobsGet"];
        put?: never;
        /** RssGetPickupJobs */
        post: operations["postRssPickupjobsGet"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/rss/pickupLocations/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** RssGetPickupLocations */
        get: operations["getRssPickupLocationsGet"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/rss/product/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** RssGetProduct */
        get: operations["getRssProductGet"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/rss/products/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** RssGetProducts */
        get: operations["getRssProductsGet"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/rss/stockLot/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** RssGetStockLot */
        get: operations["getRssStockLotGet"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/rss/stockLots/get": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** RssGetStockLots */
        get: operations["getRssStockLotsGet"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/rss/stockLot/update": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** RssUpdateStockLot */
        get: operations["getRssStockLotUpdate"];
        put?: never;
        /** RssUpdateStockLot */
        post: operations["postRssStockLotUpdate"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}